Worked out the system with Pytest fixtures to run the file conversions on test rather than run on main
